(Refact) seperate uploading data to presignedUrl login from submitProfileUpdate.ts
